refactor: extract Dashboard hooks, add ErrorBoundary
Dashboard.tsx (898→~300 lines) split into:
- useConnectionsManager: connection state, filtering, caching
- useMessageGeneration: workflow state, approval flow, message history
- useLinkedInSearch: search trigger, loading state integration

Added:
- ErrorBoundary.tsx: generic error boundary with retry button
- Barrel exports in features/connections, messages, search
