+++ frontend/src/features/search/index.ts
@@ -5,3 +5,4 @@ export { default as ResearchResultsCard } from './components/ResearchResultsCard
 
 // Hooks
 export { default as useSearchResults } from './hooks/useSearchResults';
+export { useLinkedInSearch } from './hooks/useLinkedInSearch';
