+++ frontend/src/shared/components/ErrorBoundary.tsx
@@ -1,0 +1,60 @@
+import { Component, type ReactNode } from 'react';
+import { AlertCircle } from 'lucide-react';
+import { Button } from '@/components/ui/button';
+
+interface ErrorBoundaryProps {
+  children: ReactNode;
+  fallbackMessage?: string;
+}
+
+interface ErrorBoundaryState {
+  hasError: boolean;
+  error: Error | null;
+}
+
+export class ErrorBoundary extends Component<ErrorBoundaryProps, ErrorBoundaryState> {
+  constructor(props: ErrorBoundaryProps) {
+    super(props);
+    this.state = { hasError: false, error: null };
+  }
+
+  static getDerivedStateFromError(error: Error): ErrorBoundaryState {
+    return { hasError: true, error };
+  }
+
+  componentDidCatch(error: Error, errorInfo: React.ErrorInfo) {
+    console.error('ErrorBoundary caught an error:', error, errorInfo);
+  }
+
+  handleRetry = () => {
+    this.setState({ hasError: false, error: null });
+  };
+
+  render() {
+    if (this.state.hasError) {
+      return (
+        <div className="bg-red-500/10 border border-red-500/20 rounded-lg p-6">
+          <div className="flex items-center space-x-3">
+            <AlertCircle className="h-5 w-5 text-red-400 flex-shrink-0" />
+            <div>
+              <h3 className="text-red-300 font-medium">Something went wrong</h3>
+              <p className="text-red-400 text-sm mt-1">
+                {this.props.fallbackMessage || 'An unexpected error occurred. Please try again.'}
+              </p>
+              <Button
+                variant="outline"
+                size="sm"
+                className="mt-3 border-red-500/30 text-red-300 hover:bg-red-500/10"
+                onClick={this.handleRetry}
+              >
+                Try Again
+              </Button>
+            </div>
+          </div>
+        </div>
+      );
+    }
+
+    return this.props.children;
+  }
+}
